Added function to allow for manual update of PDP data
I added a function, public sychronized void updatePDPData(), that allows
the manual updating of the data contained in the PDP's data table. This
way the data can be updated even though the thread is "stopped". This
function will have to be explicitly called by the implementing code. It
is recommended to check if the thread is stopped before calling this
function.